pyproject.toml
[tool.poetry]
name = "keycloak-test"
version = "0.1.0"
description = ""
authors = ["Cristian Rasch <cristianrasch@fastmail.fm>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
python-dotenv = "^0.19.2"
oidcrp = "^2.1.3"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
watchdog = "^2.1.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
